requirements.txt
aenum==3.0.0
astroid==2.4.2
canopen==1.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
fett==0.3.2
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
networkx==2.5
pylint==2.6.0
pymongo==3.11.3
pyroute2==0.5.14
python-can==3.3.4
python-jsonrpc-server==0.3.4
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
snooty-lextudio==1.8.10.dev0
toml==0.10.2
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.26.3
watchdog==1.0.2
wrapt==1.12.1
